SGS: add alternative method to detect remote identifier
This method is used when the server doesn't issue a redirection
to the request summary page (ie: HTTP error 500).

This method assumes that the request was created successfully, and
if it can't determine the remote identifier, it will return a request
without this piece of information but with non null status and
creation date.

Change-Id: I1f230d9a719f4851f6d0f6412d0311e163dc05ff